requirements.txt
langchain-openai==0.0.8
langchain==0.1.12
python-dotenv==1.0.1
beautifulsoup4==4.12.3
llama-index==0.10.19
llama-index-embeddings-azure-openai==0.1.6
llama-index-llms-langchain==0.1.3